protocol: try moving core types here
