+++ protocol/local_proposer_test.go
@@ -1,3 +1,3 @@
-package caspaxos
+package protocol
 
 var _ Proposer = (*LocalProposer)(nil)
