+++ protocol/memory_acceptor_test.go
@@ -1,3 +1,3 @@
-package caspaxos
+package protocol
 
 var _ Acceptor = (*MemoryAcceptor)(nil)
